Separate DAPSocket from debugServer. Launch config attaches debugger
