+++ packages/examples/resources/debugger/Dockerfile
@@ -12,5 +12,6 @@ RUN mkdir -p /home/mlc/server/src
 
 COPY ./packages/examples/resources/debugger/package.json /home/mlc/server
 COPY ./packages/examples/src/debugger/server/debugServer.ts /home/mlc/server/src
+COPY ./packages/examples/src/debugger/server/DAPSocket.ts /home/mlc/server/src
 
 WORKDIR /home/mlc/server
